Restructured project to seperate screens and services
