+++ lib/services/auth_service.dart
@@ -1,0 +1,19 @@
+import 'package:firebase_auth/firebase_auth.dart';
+
+Future<UserCredential> signInWithGoogle() async {
+  // Create a new provider
+  GoogleAuthProvider googleProvider = GoogleAuthProvider();
+
+  // Once signed in, return the UserCredential
+  return await FirebaseAuth.instance.signInWithPopup(googleProvider);
+
+  // Or use signInWithRedirect
+  // return await FirebaseAuth.instance.signInWithRedirect(googleProvider);
+}
+
+Future<void> signOut() async {
+  await FirebaseAuth.instance.signOut();
+
+  // Optionally, sign out of other providers
+  // await GoogleSignIn().signOut();
+}
